+++ picocli-codegen/src/main/java/picocli/codegen/aot/graalvm/ReflectionConfigGenerator.java
@@ -2,6 +2,7 @@ package picocli.codegen.aot.graalvm;
 
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
+import picocli.CommandLine.IFactory;
 import picocli.CommandLine.Mixin;
 import picocli.CommandLine.Model.ArgGroupSpec;
 import picocli.CommandLine.Model.ArgSpec;
@@ -90,13 +91,21 @@ public class ReflectionConfigGenerator {
         @Parameters(arity = "1..*", description = "One or more classes to generate a GraalVM ReflectionConfiguration for.")
         Class<?>[] classes = new Class<?>[0];
 
+        @CommandLine.Option(names = {"-c", "--factory"}, description = "Optionally specify the fully qualified class name of the custom factory to use to instantiate the command class. " +
+                "When omitted, the default picocli factory is used.")
+        String factoryClass;
+
         @Mixin
         OutputFileMixin outputFile = new OutputFileMixin();
 
         public Integer call() throws Exception {
+            IFactory factory = CommandLine.defaultFactory();
+            if (factoryClass != null) {
+                factory = (IFactory) factory.create(Class.forName(factoryClass));
+            }
             List<CommandSpec> specs = new ArrayList<CommandSpec>();
             for (Class<?> cls : classes) {
-                specs.add(new CommandLine(cls).getCommandSpec());
+                specs.add(new CommandLine(cls, factory).getCommandSpec());
             }
             String result = ReflectionConfigGenerator.generateReflectionConfig(specs.toArray(new CommandSpec[0]));
             outputFile.write(result);